+++ src/Backpropagation.java
@@ -1,6 +1,6 @@
 
 public class Backpropagation {
-	public float optimise(float bias, float output,float learningRate) {//will be changed, for a better back propagation
+	public double optimise(double bias, double output,double learningRate) {//will be changed, for a better back propagation
 		return (bias > output) ? bias+learningRate : (bias < output) ? bias-learningRate : bias;
 	}
 }
